Create Sobel_operator_pyopencl.py
There are three methods used for edge detection. In the 1st kernel, naive method is utilized to calculate the convolution of Sobel operator and original image (orimg.bin). In the 2nd kernel, constant memory is used. Finally, local memory as well as constant memory are used in the 3rd kernel.