Create jpeg_compression_pyopencl.py
Besides OpenCL kernel, Python code of JPEG compression is also included in this program, however, running time of python is too slow, thus the python part is after '##'. 

In the code, DCT, quantization, inverse-quantization as well as IDCT of JPEG compression are implemented in the kernel. Tile algorithm is used in parallel computing.